Move a couple of jsx form files to typescript
Change-type: patch
